Added ring-topology based token passing mechanism to control CPU frequency w.r.t power budget.
Description: Each thread requests tokens to only one of its neighbors.
If the neighbor has enough tokens to provide, it will donate -- otherwise
the request will be propagated to the neighbor's neighbor and so on.

Token always flows in one direction always opposite to the direction of
making requests. This ensures duplication of requests and book-keeping.