sched/fair: Fix nr accounting in select_idle_cpu
select_idle_cpu() iterates through maximum of nr cpus to find if a CPU is
idle or not. The value of nr is set based on the average scan cost and
average idle time of the current CPU, but it is lower bounded by 4 CPUs.
Hence atleast 4 CPUs will need to be searched in the case when current CPU
is too busy.

However, due to pre-decrement check of nr makes search for only 3 CPUs in
most busiest case. Hence fix this by changing this to post-decrement.

Signed-off-by: Parth Shah <parth@linux.ibm.com>