sched/fair: Use valid target CPU for task wakeup
select_task_rq_fair() uses prev_cpu to search find best CPU a task can be
woken up on. However, change in allowed cpumask pointer may change the
behavior as prev_cpu may not be allowed for the task.

So check if the cpu is allowed by the task and if not then fix it to use
first CPU from the allowed CPUs mask.

Signed-off-by: Parth Shah <parth@linux.ibm.com>