sched/fair: Classify and prefer a CPU in polling state during task wakeup
Introduce a new method to classify if a CPU is in polling state, which can
be used to wakeup a task on. Since the polling CPU can be woken up quickly,
it is more preferable to use a CPU which are in polling state over the one
which are in deep idle states.

So, make select_idle_cpu() to prefer a CPU in polling state over an idle
CPU.

Signed-off-by: Parth Shah <parth@linux.ibm.com>