TokenSmart: Add a tunable to define MIPS threshold
MIPS threshold defines the value of classifying a frequency benefitting CPU
from others. If a workload running on a CPU has Instructions completed per
Cycles, aka IPC = X, then with increase of 100MHz, there the MIPS value
also increases by the value of X*100.

But should such workload be classified as the one getting benefit of higher
frequency?
`mips_threshold` is tunable for the IPC, such that if
- mips_threshold = 0.5, then increase in 100MHz should atleast increase
  MIPS by 50Kilo.
- mips_threshold = 1, then increase in 100MHz should atleast increase MIPS
  by 100Kilo.
else that workload/CPU is considered to be not getting benefit of that
extra 100MHz higher frequency.